Move button component to stable directory and build
