+++ src/experimental/stories/dropdown.stories.js
@@ -1,7 +1,7 @@
 import { html } from 'lit-html';
 import '../components/molecules/Dropdown/cod-dropdown';
 import '../components/atoms/DropdownMenu/cod-dropdown-menu';
-import '../components/atoms/Button/cod-button';
+import '../../stable/components/Button/cod-button';
 
 export default {
   tags: ['experimental'],
